Keep workspace provider registration worker-neutral
