+++ apps/os2/src/domains/workspaces/entrypoints/workspace-capability.ts
@@ -1,13 +1,16 @@
 import { WorkerEntrypoint } from "cloudflare:workers";
 import { getInitializedDoStub } from "@iterate-com/shared/durable-object-utils/mixins/with-lifecycle-hooks";
 import type { ExecuteCodemodeFunctionCallInput } from "@iterate-com/shared/stream-processors/codemode/implementation";
-import type { ToolProviderRegistration } from "@iterate-com/shared/stream-processors/codemode/contract";
 import {
   getWorkspaceDurableObjectName,
   type CloudflareShellState,
   type WorkspaceDurableObject,
   type WorkspaceStructuredName,
 } from "~/domains/workspaces/durable-objects/workspace-durable-object.ts";
+export {
+  createWorkspaceProviderRegistration,
+  defaultWorkspaceIdForCodemodeSession,
+} from "~/domains/workspaces/entrypoints/workspace-provider-registration.ts";
 
 type WorkspaceCapabilityEnv = {
   WORKSPACE?: DurableObjectNamespace<WorkspaceDurableObject>;
@@ -91,40 +94,6 @@ export class WorkspaceCapability extends WorkerEntrypoint<
     await request?.callback?.(payload);
     return payload;
   }
-}
-
-export function createWorkspaceProviderRegistration(input: {
-  projectId: string;
-  streamPath: string;
-  workspaceId?: string;
-}): ToolProviderRegistration {
-  const workspaceId = input.workspaceId ?? defaultWorkspaceIdForCodemodeSession(input);
-  return {
-    path: ["workspace"],
-    instructions:
-      "Use ctx.workspace.<state method>(...) for durable workspace files, for example ctx.workspace.readFile(path), ctx.workspace.writeFile(path, content), ctx.workspace.applyEdits(edits). Use ctx.workspace.git.<command>(...) for Git commands, for example ctx.workspace.git.clone({ url, dir, depth, token }), ctx.workspace.git.add({ filepath, dir }), ctx.workspace.git.commit({ message, dir }), and ctx.workspace.git.push({ remote, ref, dir, token }).",
-    invocation: {
-      kind: "rpc",
-      callable: {
-        type: "workers-rpc",
-        via: {
-          type: "loopback-binding",
-          bindingType: "service",
-          exportName: "WorkspaceCapability",
-          props: {
-            projectId: input.projectId,
-            workspaceId,
-          },
-        },
-        rpcMethod: "executeCodemodeFunctionCall",
-        argsMode: "object",
-      },
-    },
-  };
-}
-
-export function defaultWorkspaceIdForCodemodeSession(input: { streamPath: string }) {
-  return `codemode-session:${input.streamPath}`;
 }
 
 async function callMethod(input: {
